+++ src/main/gen/org/jetbrains/lama/psi/impl/LamaSyntaxPrimaryExpressionImpl.java
@@ -10,27 +10,20 @@ import com.intellij.psi.util.PsiTreeUtil;
 import static org.jetbrains.lama.parser.LamaElementTypes.*;
 import org.jetbrains.lama.psi.api.*;
 
-public class LamaSyntaxPrimaryArrayImpl extends LamaSyntaxPrimaryExpressionImpl implements LamaSyntaxPrimaryArray {
+public class LamaSyntaxPrimaryExpressionImpl extends LamaElementImpl implements LamaSyntaxPrimaryExpression {
 
-  public LamaSyntaxPrimaryArrayImpl(@NotNull ASTNode node) {
+  public LamaSyntaxPrimaryExpressionImpl(@NotNull ASTNode node) {
     super(node);
   }
 
-  @Override
   public void accept(@NotNull LamaVisitor visitor) {
-    visitor.visitSyntaxPrimaryArray(this);
+    visitor.visitSyntaxPrimaryExpression(this);
   }
 
   @Override
   public void accept(@NotNull PsiElementVisitor visitor) {
     if (visitor instanceof LamaVisitor) accept((LamaVisitor)visitor);
     else super.accept(visitor);
-  }
-
-  @Override
-  @NotNull
-  public List<LamaExpressionSeries> getExpressionSeriesList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, LamaExpressionSeries.class);
   }
 
 }